[CONTENT] unwrap po files (#4929)
In https://github.com/hedyorg/hedy/pull/4767 we have unwrapped all po files hoping to never ever again have annoying issues. It went well for a while, but.... for reasons I can't understand, #4878 again wrapped the po files :( 

<img width="1470" alt="image" src="https://github.com/hedyorg/hedy/assets/1003685/e7375971-a41f-4746-ba94-b230a6d524ae">

This unwraps everything

Additionally, @rix0rrr created #4930 which adds a test to test for accidental unwrapping. That will not solve the issue but will 1) prevent issues and 2) might help us to understand when it goes wrong and with that why.