[CHORE] Check .po files for accidental wrapping (#4930)
Translations in the `.po` files should never be wrapped. When we extract them, we configure `--no-wrap`, and we also configure Weblate to never wrap them.

However, for a reason we don't understand, wrapping still occurred in https://github.com/hedyorg/hedy/pull/4878.

Add a test to detect this so we don't run into the problem again.