[FIX] Remove fixed height for editor (#4915)
Fixes #4895

We used to set fixed pixel heights for the (CodeMirror) editor; one of the things we would do is measure the height of the container during editor initialization, and then force the editor to have that height.

On the slides page, where we would load a lot of pages in `<iframe>`s that would start out hidden, the calculated height would sometimes be measured as `0px`, which would then be copied onto the editor. The end result would be that the editor would appear empty on some pages.

Instead, in this PR: remove all height calculations. We will set a height on the grid cells that contain the editor and output box, and add the CSS declarations necessary to make the actual editor automatically resize to take up available remaining space, after any fixed elements in the same cell are taken away (like the error and success popups, the "handed in" warning box, etc).

This makes it no longer necessary to recalculate the height of the editor every time one of those popups is shown or hidden.

For the "developer mode" toggle, we introduce a new attribute: `data-devmodeheight="22rem,36rem"`, and the toggle will switch between the two sizes in the attribute. This avoids having to duplicate element sizes between HTML and JavaScript.

**How to test**

Verify this solves #4895 by examining the slides of level 1 and seeing that all code blocks are always shown.