Embedded editor improvements (#4936)
**Description**
In this PR we make some improvements to the embedded editor. The editor itself was updated recently to be better responsive, but the embedded editor was left out-of-scope. We make some improvements to the scale-ability, fixing some overflow with long lines of code and automatically grow the editor to the available height.

**Fixes**
This has no related issue

**Example**
**Old**
<img width="1512" alt="Scherm­afbeelding 2023-12-27 om 10 02 37" src="https://github.com/hedyorg/hedy/assets/38155424/43e3c7ba-1933-4a32-90dd-e7086a368f04">
**Fixed**
<img width="1512" alt="Scherm­afbeelding 2023-12-27 om 10 02 44" src="https://github.com/hedyorg/hedy/assets/38155424/ee4f972f-5eeb-46c9-a3af-dbfb9a313c7a">


**How to test**
http://127.0.0.1:8080/embedded/1/?lang=nl&run=true&readOnly=false&program=cHJpbnQgSG9pLCB3ZWxrb20gYmlqIEhlZHkhIEhvaSwgd2Vsa29tIGJpaiBIZWR5ISBIb2ksIHdlbGtvbSBiaWogSGVkeSEgSG9pLCB3ZWxrb20gYmlqIEhlZHkhIEhvaSwgd2Vsa29tIGJpaiBIZWR5IQphc2sgSG9lIGdhYXQgaGV0IG1ldCBqZT8KXyBEdXMgaGV0IGdhYXQ=