[UI] Switch read only editors to code mirror (#4924)
**Description**

We are currently using CodeMirror in the main editor, and Ace in the read-only editors, this situation is not ideal since we need to maintain two wildly different highlighting systems, therefore this PR switches all the editors to use CodeMirror.

This is the current state of the replacement:

* [x] Example editors in the level page
* [x] Editors in the quiz's questions
* [x] Parsons
* [x] Explore page editors
* [x] Create your own adventure page's editors
* [x] Common mistakes page of the teacher's manual
* [x] Editors in the Cheatsheet page (fixes #4935)

**About Parsons**

The parsons exercises where deeply coupled with Ace, and the dragging code was difficult to understand. In order to switch these editors to CodeMirror it was necessary a redo of most of the code, therefore, this PR also fixes #4442 because we are now using the Sortable library, making the dragging easier. This is the parson interface now:
![image](https://github.com/hedyorg/hedy/assets/45865185/305b93ea-1b8e-40a0-bdd8-1302b3168a08)

Now students can simply shuffle the lines in place, instead of dragging them from one place to the other.