[REFACTOR] Rename `start_code` to `editor_contents`
We're not using `start_code` anymore for Adventures, but there is still
a field called `start_code` in the back-end to front-end communication.

Rename that field to `editor_contents` to more clearly indicate its
purpose.

Finally remove `start_code` from a couple of places where it explicitly
meant the `start_code` in the YAML.