🧹 Rename `start_code` to `editor_contents` (#4942)
We're not using `start_code` anymore for Adventures, but there is still a field called `start_code` in the back-end to front-end communication.

Rename that field to `editor_contents` to more clearly indicate its purpose.

Finally remove `start_code` from a couple of places where it explicitly meant the `start_code` in the YAML.

This change also made `editor_contents` have a default value so it doesn't need to be passed when instantiating `Adventure()`. That requires making the fields after it have defaults as well, or moving the field around. Because there was an instantiating of `Adventure('start', 'start', 'start', 'start', code, False, False)`, changing the field order was too dangerous. That code probably shouldn't have been allowed in the first place, so also add a decorator to make it required to pass all fields as keyword arguments (instantiation will fail if positional arguments are used).

Fixes #4922

**How to test**

This PR did not make any functional changes. Everything should still work as before.