[BUG FIX] Fix embedded editor when in read only mode (#4952)
This PR fixes a bug where there would be whitespace add the bottom of the embedded editor when on readonly and disabled run mode. This because we calculated the height of the editor based on the height of the run button. But if it's not there, no need to take this space!

The following test link should no longer have whitespace at the bottom:
http://127.0.0.1:8080/embedded/1/?lang=nl&run=false&readOnly=true&program=cHJpbnQgZGl0IGlzIGVlbiB0ZXN0IQplY2hvIHdhYXJvbSBrbG9wdCBkaXQgbmlldD8KdGVzdCB0ZXN0CnRlc3QgdGVzdA%3D%3D