🧹 Add a build tool (#4799)
We are accumulating quite a lot of build scripts that are all written in bash, and need to be implemented in a just-so order, and you need to know exactly when you need to run them. It might be time to start looking into moving to an actual tool to manage the build for us.

These come in a lot of flavors with various capabilities. I've looked around a bunch, and I think [DoIt](https://pydoit.org) looks quite nice.

Here are the features that I think make this tool quite nice:

- It's written in Python, so comes in automatically as Python dependency during `pip install`.
- The configuration file is written in Python as well, so we don't have to learn a new configuration language just for this tool.
- Tasks have dependencies between them, so if there are multiple that need to run, they automatically run in the right order.
- Tasks can (optionally) declare files that are inputs to the task itself, and if the input files have not changed the task is skipped. This makes it safe to just run the build toold and trust that you're not waiting too long for nothing.

I've started converting some build scripts over to this new tool to show how it works. You can see it in action by going into the virtual env and running (must be in the root directory, unfortunately):

```
$ doit
$ doit run typescript
$ doit help
```

The `.doit.db.db` that I've ignored is a file it uses internally to track changes to dependencies.

I've paused here because I'd like to solicit feedback.  Does the tool look okay? Is this understandable? Do we prefer this to bash, or to other build tools?