📜 quiz for level17 (#4839)
I have made a preliminary quiz for level 17, added for now only four questions. This PR is just to make sure I am on the right track but also if I am allowed to continue working on this. It took me a bit of time to figure out where I can find the quizzes and how I can write my tests, but now I'm on a roll.