📜 Re-add exclamation mark in question in quiz level 2 (#5082)
Fixes #5068 

**How to test**

Follow these steps to verify this PR works as intended:

Go to level 2 and see if in the Quiz question 4 does not have an '!' in the example code anymore, in the English and the Dutch version.


**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!