🚚 Prevent failure page after logout
We try to keep the user on the same page after they log out, to not
disturb their place on the website. This leads to the experience that
when you log out and you are on a page that was only visible if you are
logged in, you will see an error message.

Add a fix that the first page load that fails due to permissions
after a logout is turned into a redirect to the front page. That way,
if you were on an unprivileged page you stay where you are, but if you
were on a privileged page you will not see an error.