📜 Curlies for english  (#5171)
@Felienne: 

I have cleaned up the curlies from level 18 up to level 14 (all tabs including quiz). 

At level 14 by the "is" tab I am not sure what part of the code has to be where. See screenshots, the "localhost url" is what I made, the other one is how it is looks in reality.

<img width="1353" alt="whatimade" src="https://github.com/hedyorg/hedy/assets/127490004/0218246b-3bc9-4bd6-8fe1-de89f25e906d">
<img width="1368" alt="whatisnow" src="https://github.com/hedyorg/hedy/assets/127490004/7ac1cc07-3909-43cf-b886-69d768e8c356">