🖊️ Allow -= and > operators to be used independently when merging grammars (#5186)
Fixes #5176

Merging grammar rules was only possible if the += operator was present, which means that the -= and > operators could not be used independently of +=. With this change, all operators could be used independently of each other.

Because the > operator now does not need the context of the += operator, I had to change it to a symbol which does not appear in the lark language (e.g. ->). Please consider whether there is a better candidate than `>>`, for example `>=`?

**How to test**

- The introduced merging changes are covered with unit tests
- All existing tests should pass