🪲 Fix old programs in program page, improve filtering teacher programs (#5197)
Same thing happened to us back in #4417. Some old programs don't have the field `adventure_name` in them, and thus our systems fails. I also improve a bit the filtering of programs, by not using the program name, but rather the adventure name.

I also have a question, what to do with programs that might have an id that's neither in the DB or our own contents? Right now, Im resorting to just show the id, but perhaps we should skip them? What do you think?

![image](https://github.com/hedyorg/hedy/assets/45865185/dd65b538-960d-40da-a02f-a2f4336daa7d)


**How to test**

Go to My Programs page, it should be working. 

If you want to see a program with a non-existing adventure linked, copy an existing `program` record in the dev_dabase.json and change the `adventure_name` field to something that's not present in the database.