🚚 Don't rewrite emoji in YAML files
The current autoformatting script is replacing emoji with their
Unicode escape counterpart. Update the version of Ruamel YAML to
a version which doesn't have this behavior anymore, and keeps the
emoji in the file.

Move the running of the YAML autoformatter out of the pre-commit
job and into the doit script `_autopr` which runs in a separate
GitHub Actions job.