🚚 Squash merge Weblate PRs
We used to regular-merge Weblate PRs, which leads to a giant history
of Weblate commits in the commit history, making it hard to find
functional changes.

We weren't able to squash-merge Weblate PRs because it would easily
lead to merge conflicts.

Now that we control when Weblate pulls and creates commits, this
problem has disappeared, so we can safely squash-merge and get a nicer
history.