🚚 Prevent the endless update train when fix script reports errors (#5302)
The problem was that the script would write to a `.txt` file, which would subsequently get committed and trigger another update which would post again, etc.

Fixed by writing to `.md.tmp` files, and gitignoring `.tmp` files. While we're at it, make the comment a bit nicer to read as well by outputting MarkDown instead of plain text.