Trigger a Weblate reset after pushing translations (#5347)
After we have told Weblate to push all of its pending commits, it should forget about them.

The reason is that we may revert some code samples, and when the translations PR finally squash-merges its way onto `main` and gets back to Weblate, it will see a conflict with its own pending changes.

By resetting, it will forget about its own pending changes, and consume the changes when they finally come in from the upstream repo.

This probably leads to an inconsistency window in which it looks for about ~1hr like the changes have disappeared. Fear not, they haven't, they will come back once the PR is in and Weblate syncs from the main repo.