🪲 Fix accidental object sharing in Dynamo layer
As discovered in #5350 by @jpelay, the Dynamo emulation layer has
a bug that can lead to accidental data sharing between different
requests, because the DB layer hands out references to long-lived
objects. If those get mutated in-place by a reader, those mutations
accidentally affect the stored in-memory copy of the database.