🪲 duplicate adventures (#5385)
Fixes #5375

This PR changes how adventures are created. We don't need to "store" the adventures unless the teacher appllies some changes. This means whenever we redirect the teacher to the customize-adventure page, we simply initialize an adventrue, save in the current session, and if they apply a change, then we basically save it in our DB. 

Not only does this prevent the issue I'm fixing, but also reduce the number of trial adventures created by teachers. 

Moreover, the only downside is that teachers would have to write out the name themselves, however, this may not be a min-point afterall since giving default names is kinda encouraging creating more, and more will be created just because it becomes easy to do so!

**How to test?**
Create an adventure from the 
- for-teachers page
- class-overview page
- customize-class page