📜 Hangman adventure (#5333)
@Felienne I've made hangman the same way as the Blackjack adventure and split it up in 3 parts.

There is quite a lot of text, but without it I think it would be too hard to make this code. And it is almost the last level, so this is meant for the students who've made it this far anyway, so they could use a challenge. Could you (or someone else) test it to see if the instructions are clear enough to follow and create a game of hangman (especially part 1 and 2, you don't need to figure out how to draw tha gallow for part 3 haha, unless you feel like it) ? 

If it's doable, I will add this adventure to other languages as well and translate it to dutch.