💻 autosave feature (#5311)
Fixes #4888 

**Note**
The `autoSave` function that i created can be used generically. So, in case we need it in the future, these are the steps that are needed:
- make sure that in the HTML you have a form element with a onsubmit event (or htmx!).
   - all the inputs elements (incl. select, textarea etc) are then handled by the function
- then add a script that invokes the function with the id of that form element, as in https://github.com/hedyorg/hedy/blob/autosave/templates/profile.html#L184-L189.
- the fuction that you use to make the api request will be invoked once a change has occurred.

Furthermore, sometimes there's an element that we cannot attach an event listener on, like the CKEditor. In this case, you can pass an event as a secon argument to the `autoSave` function. When you do this, we simply invoke the submit trigger on the form element. 

The customize-class page has two differently implemented features. The sortable adventures and the other inputs below. These two have to be syncronized. If we hide all quizzes, for instance, we want the adventure tabs to be updated. Now, instead of changing the whole structure, a third argument could be passed indicating that we should trigger an event on a certain element. See the autosave.ts file on how it's used.

**How to test?**
check the description of #4888 and test these pages.