🚚 Fix Weblate merge conflict resolution script
The Weblate merge conflict resolution script begins with normalizing
the YAML files on Weblate's branch before merging `main` (which is
always normalized).

However, because we left out `--allow-empty`, it would fail if there
was nothing to normalize.

Allow creating an empty commit if there's nothing to normalize, the
empty commit will be squashed away anyway.