🚚 Prevent Weblate conflicts by locking/unlocking
We're still sometimes seeing merge conflicts in Weblate because
people are submitting translations in the ~1-2hr window it takes us to
merge the update PR.

To completely avoid these, we're going to take rough measures: lock
the Weblate repo when we start the merge process, and only unlock
it after the PR is merged.

Because merging a PR is an asynchronous process, this lock/unlock behavior
is split over 2 different workflows (`update-weblate.yml` to lock,
`unlock-weblate.yml` to unlock).

For the cases where there are still conflicts due to manual editing
in translated YAMLs, we do the same for the "Resolve conflicts" script:
make it work the same as the regular daily update script (lock + make
Weblate forget about changes it had that are now in a normalized PR).